# Speed up addHoverClass on large stylesheets