Multi-megabyte stylesheets were incredibly slow with addHoverClass. I was able to bring that back by >90% on large stylesheets at a negligible cost on tiny stylesheets.

Results included below:
```
Running "4.1M stylesheet" suite...

  existingAddHoverClass:
    0 ops/s, ±0.90%   | slowest, 100% slower

  newAddHoverClass:
    2 ops/s, ±2.98%   | fastest

Finished 2 cases!
  Fastest: newAddHoverClass
  Slowest: existingAddHoverClass
```
```
Running "248K stylesheet" suite...

  existingAddHoverClass:
    56 ops/s, ±2.10%   | slowest, 32.53% slower

  newAddHoverClass:
    83 ops/s, ±1.86%   | fastest

Finished 2 cases!
  Fastest: newAddHoverClass
  Slowest: existingAddHoverClass
```
```
Running "105B stylesheet" suite...

  existingAddHoverClass:
    22 031 ops/s, ±2.32%   | fastest

  newAddHoverClass:
    21 493 ops/s, ±3.22%   | slowest, 2.44% slower

Finished 2 cases!
  Fastest: existingAddHoverClass
  Slowest: newAddHoverClass
```
Benchmark ran in node.js on a MacBook Pro (16-inch, 2019), 2,3 GHz 8-Core Intel Core i9.
Similar results found when running tests in Safari and Chrome.
[Visual charts of these benchmarks: `results.zip`](https://github.com/rrweb-io/rrweb-snapshot/files/6303786/results.zip)