@Juice10 2021-04-14
I ran into a bug in one of the integration tests so I came up with a couple new variants. `addHoverClassFilterFirst` seemed to have the most upside with the least downside so I ended up picking that one: It is only 5% slower on the smallest stylesheets and it's still 99% faster on the larger stylesheets.

## Benchmarks
[Download graphs and granular results.](https://github.com/rrweb-io/rrweb-snapshot/files/6308669/Archive.zip)

### 4.1M stylesheet
```
  addHoverClassOriginal:
    0.02 ops/s, ±0.80%   | slowest, 99.04% slower

  addHoverClassFilterFirst:
    2.09 ops/s, ±4.93%   | fastest

  addHoverClassSortFirst:
    2.07 ops/s, ±4.88%   | 0.96% slower

  addHoverClassUnique:
    1.97 ops/s, ±4.90%   | 5.74% slower

Finished 4 cases!
  Fastest: addHoverClassFilterFirst
  Slowest: addHoverClassOriginal
```

### 248K stylesheet
```
  addHoverClassOriginal:
    56.3 ops/s, ±1.01%   | slowest, 33.45% slower

  addHoverClassFilterFirst:
    83.8 ops/s, ±1.67%   | 0.95% slower

  addHoverClassSortFirst:
    84.6 ops/s, ±1.09%   | fastest

  addHoverClassUnique:
    84.3 ops/s, ±0.72%   | 0.35% slower

Finished 4 cases!
  Fastest: addHoverClassSortFirst
  Slowest: addHoverClassOriginal
```

### 105B stylesheet
```
  addHoverClassOriginal:
    23 851 ops/s, ±2.20%   | fastest

  addHoverClassFilterFirst:
    22 672 ops/s, ±2.97%   | 4.94% slower

  addHoverClassSortFirst:
    22 322 ops/s, ±2.10%   | slowest, 6.41% slower

  addHoverClassUnique:
    22 339 ops/s, ±2.89%   | 6.34% slower

Finished 4 cases!
  Fastest: addHoverClassOriginal
  Slowest: addHoverClassSortFirst
```

## Tests pass for each variant
```
  add hover class to hover selector related rules
    addHoverClassOriginal
      ✓ can add hover class when there is a multi selector with the same prefix
      ✓ will do nothing to css text without :hover
      ✓ can add hover class to css text
      ✓ can add hover class when there is multi selector
      ✓ can add hover class when :hover is not the end of selector
      ✓ can add hover class when the selector has multi :hover
      ✓ will ignore :hover in css value
    addHoverClassFilterFirst
      ✓ can add hover class when there is a multi selector with the same prefix
      ✓ will do nothing to css text without :hover
      ✓ can add hover class to css text
      ✓ can add hover class when there is multi selector
      ✓ can add hover class when :hover is not the end of selector
      ✓ can add hover class when the selector has multi :hover
      ✓ will ignore :hover in css value
    addHoverClassSortFirst
      ✓ can add hover class when there is a multi selector with the same prefix
      ✓ will do nothing to css text without :hover
      ✓ can add hover class to css text
      ✓ can add hover class when there is multi selector
      ✓ can add hover class when :hover is not the end of selector
      ✓ can add hover class when the selector has multi :hover
      ✓ will ignore :hover in css value
    addHoverClassUnique
      ✓ can add hover class when there is a multi selector with the same prefix
      ✓ will do nothing to css text without :hover
      ✓ can add hover class to css text
      ✓ can add hover class when there is multi selector
      ✓ can add hover class when :hover is not the end of selector
      ✓ can add hover class when the selector has multi :hover
      ✓ will ignore :hover in css value


  28 passing (8m)
```

## Different test cases:
### addHoverClassFilterFirst
```js
const HOVER_SELECTOR = /([^\\]):hover/;
const HOVER_SELECTOR_GLOBAL = new RegExp(HOVER_SELECTOR, 'g');
export function addHoverClassFilterFirst(cssText: string): string {
  const ast = parse(cssText, {
    silent: true,
  });

  if (!ast.stylesheet) {
    return cssText;
  }

  const selectors: string[] = [];
  ast.stylesheet.rules.forEach((rule) => {
    if ('selectors' in rule) {
      (rule.selectors || []).forEach((selector: string) => {
        if (HOVER_SELECTOR.test(selector)) {
          selectors.push(selector);
        }
      });
    }
  });

  if (selectors.length === 0) return cssText;

  const selectorMatcher = new RegExp(
    selectors
      .filter((selector, index) => selectors.indexOf(selector) === index)
      .sort((a, b) => b.length - a.length)
      .map((selector) => {
        return escapeRegExp(selector);
      })
      .join('|'),
    'g',
  );

  return cssText.replace(selectorMatcher, (selector) => {
    const newSelector = selector.replace(HOVER_SELECTOR_GLOBAL, '$1.\\:hover');
    return `${selector}, ${newSelector}`;
  });
}
```
### addHoverClassSortFirst
```js
export function addHoverClassSortFirst(cssText: string): string {
  const ast = parse(cssText, {
    silent: true,
  });

  if (!ast.stylesheet) {
    return cssText;
  }

  const selectors: string[] = [];
  ast.stylesheet.rules.forEach((rule) => {
    if ('selectors' in rule) {
      (rule.selectors || []).forEach((selector: string) => {
        if (HOVER_SELECTOR.test(selector)) {
          selectors.push(selector);
        }
      });
    }
  });

  if (selectors.length === 0) return cssText;

  const selectorMatcher = new RegExp(
    selectors
      .sort((a, b) => b.length - a.length)
      .filter((selector, index) => selectors.indexOf(selector) === index)
      .map((selector) => {
        return escapeRegExp(selector);
      })
      .join('|'),
    'g',
  );

  return cssText.replace(selectorMatcher, (selector) => {
    const newSelector = selector.replace(HOVER_SELECTOR_GLOBAL, '$1.\\:hover');
    return `${selector}, ${newSelector}`;
  });
}
```
### addHoverClassUnique
```js
function unique(arr: string[]): string[] {
  if (!Array.isArray(arr)) {
    throw new TypeError('array-unique expects an array.');
  }

  const len = arr.length;
  let i = -1;

  while (i++ < len) {
    let j = i + 1;

    for (; j < arr.length; ++j) {
      if (arr[i] === arr[j]) {
        arr.splice(j--, 1);
      }
    }
  }
  return arr;
}

export function addHoverClassUnique(cssText: string): string {
  const ast = parse(cssText, {
    silent: true,
  });

  if (!ast.stylesheet) {
    return cssText;
  }

  const selectors: string[] = [];
  ast.stylesheet.rules.forEach((rule) => {
    if ('selectors' in rule) {
      (rule.selectors || []).forEach((selector: string) => {
        if (HOVER_SELECTOR.test(selector)) {
          selectors.push(selector);
        }
      });
    }
  });

  if (selectors.length === 0) return cssText;

  const selectorMatcher = new RegExp(
    unique(selectors)
      .sort((a, b) => b.length - a.length)
      // .filter((selector, index) => selectors.indexOf(selector) === index)
      .map((selector) => {
        return escapeRegExp(selector);
      })
      .join('|'),
    'g',
  );

  return cssText.replace(selectorMatcher, (selector) => {
    const newSelector = selector.replace(HOVER_SELECTOR_GLOBAL, '$1.\\:hover');
    return `${selector}, ${newSelector}`;
  });
}
```