@Yuyz0112 2021-04-22
Thanks, @Juice10! The implementation looks great to me and I can reproduce similar results in some stylesheet examples.